third_party/blink/web_tests/svg/dynamic-updates/SVGTextElement-svgdom-transform-prop-expected.txt

SVG 1.1 dynamic update tests
Text content

Tests dynamic updates of the 'transform' property of the SVGTextElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textElement.transform.baseVal.getItem(0).matrix.f is 0.0
PASS textElement.transform.baseVal.getItem(0).matrix.f is -200.0
PASS successfullyParsed is true

TEST COMPLETE
